Extracted Renderer Route Extension
