+++ src/Framework/Template/Php/Extension.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Framework\Template\Php;
+
+abstract class Extension
+{
+    public function getFunctions(): array
+    {
+        return [];
+    }
+}
