+++ src/Framework/di/dependencies.php
@@ -7,7 +7,8 @@ use Framework\Http\ActionResolver;
 use Framework\Http\HttpApplication;
 use Framework\Http\MiddlewareResolver;
 use Framework\Http\Router\Router;
-use Framework\Template\PhpRenderer;
+use Framework\Template\Php\Extension\RouteExtension;
+use Framework\Template\Php\PhpRenderer;
 use Laminas\Diactoros\Response;
 
 return [
@@ -25,6 +26,8 @@ return [
     }),
 
     TemplateRenderer::class => DependencyInjection\factory(function (ApplicationInterface $container) {
-        return new PhpRenderer('templates', $container->get(Router::class));
+        $renderer = new PhpRenderer('templates');
+        $renderer->addExtension($container->get(RouteExtension::class));
+        return $renderer;
     }),
 ];
